PR changes and other minnor changes

+++ include/dds/core/xtypes/detail/PrimitiveTypes.hpp
@@ -42,21 +42,21 @@ template<>
 struct dynamic_type_traits<int16_t>
 {
     static constexpr TypeKind TYPE_ID = TypeKind::INT_16_TYPE;
-    static constexpr const char* NAME = "uint16_t";
+    static constexpr const char* NAME = "int16_t";
 };
 
 template<>
 struct dynamic_type_traits<uint16_t>
 {
     static constexpr TypeKind TYPE_ID = TypeKind::UINT_16_TYPE;
-    static constexpr const char* NAME = "int16_t";
+    static constexpr const char* NAME = "uint16_t";
 };
 
 template<>
 struct dynamic_type_traits<int32_t>
 {
     static constexpr TypeKind TYPE_ID = TypeKind::INT_32_TYPE;
-    static constexpr const char* NAME = "int16_t";
+    static constexpr const char* NAME = "int32_t";
 };
 
 template<>
